Display view.commonwl.org visualization #1058
For CWL workflows, display an SVG created by view.commonwl.org. This
is in addition to the current DAG visualization, and the change
includes radiobuttons that let the user select which visualization
to see.

This feature is off by default, as view.commonwl.org does not
support CORS. Two issues were filed and fixed in that repo, but they
have not yet been deployed. This was tested by running an instance of
the viewer locally.

To turn on the feature, initialize enableCwlViewer to true in
src/app/workflow/dag/dag.component.ts. Ideally a more fancy feature
flag feature should have been implemented that doesn't require
changing the source code to toggle the feature, but it didn't seem worth
the effort unless this is a need we anticipate coming up more often.

Todo:

The zooming in/out is pretty crude -- it doesn't let you pick which
part of the image you to see.