Fix broken DAG #1199
It was only failing in Angular prod mode.

Using source maps, I saw it was failing in cytoscape.min.js, so then
I tried the non-minified version to see exactly what the problem was,
and the problem went away.

The problem seems to be unrelated to the CWL Viewer DAG work, as I could
still reproduce the problem in prod mode if I rolled back the CWL
Viewer commits locally.

With this change, the generated bundles only seem to contain minified
code; it looks like the difference is that Angular preforms the
minification at build time.