Add Feature Flags #268
Add a "feature flags" property in dockstore.model.ts, so that
features can be turned on/off without doing new Git commits.
EnableCwlViewer is the only flag in there so far.

Also added a comment to dockstore.model.ts warning to keep it in sync
with the template in the dockstore/compose_setup repo, as I was
previously unaware of the need.