Use secondaryFiles for FC Launch Eligibility (#234)
We were trying to not expose Launch with Firecloud if the WDL contained
any WDL import statements. The test for this was a simple Regex that
looked for any lines that began with import. This was not a good test
because there was a workflow that had some inline Python code that
was doing an import.

On further discussion with FireCloud team, FireCloud does support
http imports. So instead rely on `workflowsService.getSecondaryWdl` to
indicate whether the WDL has any *file* imports, and disable in that
case.

This change basically gutted firecloud.service.ts, so got rid of it
as well.