Fix Incomplete Source File Display (#271)
ga4gh/dockstore#1357

Problem was that source code was not escaped. Changed
FileService.highlightCode to escape. There was also a bug in that
code where it had the closing </code> tag after the closing </pre>
tag.

In researching/testing it, I noticed that `src/app/paramfiles` is not
used any more, so I deleted it as part of this PR.