Don't block CWL Viewer on classic Dag error
ga4gh/dockstore#1487

Changed HTML to not block CWL viewer. Also tried
to simplify it a little bit. Still a bit of a mess
though, as in dag.component.html, there are error
elements only for DAG, but some controls right in the
middle that also affect the CWL Viewer.

* Also removed a couple of unused fields
* And added a default error message in cwl viewer if one didn't
come back from server.