Disable instead of hide Launch with Buttons
Instead of hiding Launch with FireCloud/DNAstack buttons when the WDL
is not supported, disable them so that the user knows what is going on.

For FireCloud, disable if there are file-based imports.

For DNAstack, disable if there are any imports.

Had to change to buttons from anchors, because you cannot disable
anchors (well, not with disabled attribute).

Test for http(s) imports is sort of crude; we should probably do it
server side using... wom?.

Also, there is some code duplication in checking if there are imports
for FC and DNAstack. But since this code is going into hotfix and
develop branch is on Angular 6, decided to do it that way rather than
using RxJs6 operators (which I don't know very well, plus wouldn't work
on hotfix).