Properly Check Permissions for Shared Workflow
A fix for ga4gh/dockstore#1589

We have to go with this approach with the current webservice API,
which makes me the question of the current permissions API.

The logic is, if you can fetch permissions, you are an owner. If you
are not, then make OPTIONS calls to see what APIs you can invoke,
and the `Allow` header in the response will indicate what HTTP method
you have permissions to invoke.

My thinking was:

/path/workflow/{repository}/permissions should be an unauthorized call
if you are not an owner of {repository} -- e.g., if you only have read
permissions on the workflow, you should be able to see the workflow
but not who it's been shared with (in fact, that is how the policies
are currently defined in SAM, although we could change them).

Given that, I thought that the API could support OPTIONS calls to
indicate which endpoints are invocable with which methods. I liked
that, because if I have an endpoint that simply says for example that
you have READ permissions on a workflow, then the UI has to figure out
which API calls are allowed with that permission and which are not.
Whereas with OPTIONS you can query the API call you are going to make
and let the backend tell you whether you have permissions or not.

However:

1. I've only added a couple of OPTIONS endpoints, for the endpoints that
currently can be affected by permissions, so it's incomplete.
2. This UI code is convoluted.

So maybe we should just enable the permissions endpoint for all users,
but if you are not an owner, it only returns the current user's
permissions. That means the UI would have to figure out what API calls
it can and can't make, but it looks like it was essentially doing
that anyway.