Add Launch with DNAnexus (#352)
* Add Launch with DNAnexus

ga4gh/dockstore#1537

Added launch with DNAnexus button. The order of the launch with
buttons in the UI is chronological.

Got rid of enableLaunchWithFireCloud feature flag. It was there
because at one point we were only able to test this feature on their
side in staging but not in prod, but now that it's been in prod
for a while, I think we can get rid of the flag and the tests for it.

Also introduced a service to generate the urls. Use HttpParams there
instead of the URLSearchParams being used previously in the
component because URLSearchParams is deprecated.

Note: I'm planning to tackle ga4gh/dockstore#1575 next in the UI, which
relates to disabling launch with buttons based on the file content.
Wanted to to do this on its own first.