WIP: Better Handling for Permissions (#367)
* Better Handling for Permissions

UI portion of ga4gh/dockstore#1589

This has a dependency on PR ga4gh/dockstore#1638, which has a new
endpoint.

Call new endpoint to get allowable actions. If SHARE action is one
of the actions, then user is an owner, and go ahead and fetch all
permissions for the workflow.

Unfortunately, generated TS code doesn't generate a type for the enum
returned by Actions.