More prominent error message for SAM edge case
ga4gh/dockstore#1713

Display the red box if there is an issue where the wrong user owns
a SAM resource.

I don't think we should display the status code in the message, but
that is discussed in ga4gh/dockstore#1729 and should be handled
generically.

Also waiting on the generic solution there for removing the error
message after success (although in this scenario, I think you can
only get success if you go to another page).