Don't create 2 instances of UserService (#437)
ga4gh/dockstore#1764

UserService is listed as a provider in app.module.ts. A second instance
of the UserService was getting loaded due to the 2 declarations, in
here and in app.module.ts, which led to odd behavior:

Logging in and out out would only update one of the instances of the
service, and then when the other instance came into play for some
components, that instances would still have a reference to the previous
user.