Use progress bar. Also updated docker and param files
Things are tricky in entry-file-selector.ts with the call to
`getFiles()` in `reactToDescriptor()`. Even though it is called
just once here, call to `this.gA4GHFilesService.clearFiles()` in
entry.ts causes two observables to be sent to the subscription,
one with an empty array, and one with the version's descriptors,
which may or may not be empty. Thus there is no easy way to
see when the descriptor has loaded.

And thus there is a delay for the progress bar to appear while the
descriptors are being downloaded. But that delay is fairly small,
even at 3G, so I think it's acceptable.