Check if descriptors exist before saying they don't (#502)
ga4gh/dockstore#1937

When there are no descriptors, `nullDescriptors` is set to `true`, so go
off that instead of content.

Also added a progress bar.

Use `setError` to help determine if the store has no files because there are legitimately no files vs. thre was an error trying to fetch the files.