If webservice on port 443, omit port in url
ga4gh/dockstore#1594

This doesn't drop the port 80 if on default port of http, but
doesn't seem worth it.

There are a lot of places that use API_URL; instead of massaging
them one by one to remove the 443 port, it seems like we
should just do it here. A lot of those places have user-facing
URLs, e.g., the TRS url, the launch with instructions...

Will create a corresponding PR in compose_setup if this gets
approved.