Feature/2386/config (#628)
dockstore/dockstore#2386

All api calls go to /api  on same domain (host/port) as UI is served from endpoint -- see app.module.ts

In the APP_INITIALIZER, UI calls /api/metadata/config.json and sets dockstore.model.ts based on the response.

When running locally, use Angular proxy. Just use `npm run start`.

When running on Circle-CI, use nginx.

Also: 
In checkerWorkflowFromWorkflow.ts, it was checking
if a button was not visible, then if it was visible.
Removed that, as I think it was relying on a delay
for the button to appear.