Fix refreshing Google tokens (#746)
* Fix refreshing Google tokens

dockstore/dockstore#2745

Problem was we weren't getting a refresh token back in the
initial authentication dance.

Looks like a dependabot update of ng2-ui-auth caused this.

Changed AuthConfig in auth.model to have a type (when I changed
it, compilation errors happened), then adjusted the code, which
also included getting rid of defaultHeaders field, which isn't part
of the type.

Regarding, changing the value of prompt to "consent", this page,
https://developers.google.com/identity/protocols/OpenIDConnect#prompt,
says the only allowed values are none, consent, and select_account.
I went with consent, because you may not get a refresh token if
you already requested an access token only before. See
https://github.com/googleapis/google-api-python-client/issues/213.

Had to use TypeScript User Defined Type Guards for the first time in
configuration.service.ts to fix compilation errors.