Refresh individual repos instead of org
dockstore/dockstore#3485

When user clicks Refresh Organization button, refresh each repo
individually instead of invoking the soon to be disappearing endpoint
that refreshes the whole org.

Need access to orgEntryObject, so passing that to component instead
of org/namespace.

Did not do tools, although did some of the prep work for it.

Added new method in WorkflowService to update a workflow, sort of
confusing (to me) in that there are two Akita stores that need to get
updated.

Also changed RefreshOrganizationComponent to extend Base and eliminate
some code duplication.