add --prod flag to all npm build steps
This PR modifies package.json to add the --prod flag to the build step
for the staging stage (not just for the prod stage).

This is necessary because without the --prod flag, the staging site
adds a <noscript></noscript> tag block that contains <link> tags
that do not have an 'integrity' attribute.

In other words, without the --prod flag, users with Javascript turned
off could be vulnerable to CSS-based cross-site script attacks.

This pull request adds the --prod flag to the non-prod build stage.

See [SEAB-3225](https://ucsc-cgl.atlassian.net/browse/SEAB-3225).