Upload to s3 if audit and unit tests pass (#1502)
Also reorganized Cypress tests, splitting largest group into two groups.

https://ucsc-cgl.atlassian.net/browse/SEAB-4198

Tests would not pass until they got reorganized. The theory was that
we were hitting resource limits, and the instance was bogging down.

Splitting them up seems to be working, with a small sample size,
and is also improving the total time to run the tests.