Update artifacts; fix flaky test
The Twitter "placeholder" div doesn't exist in the new dashboard, the
lack of which would frequently cause a native element exception and a
Cypress failure.

Ideally there would be another component that's a container for the old
and new dashboards, instead of the way it is now, where this is both
the old dashboard and a container for the new dashboard. But I may want
to bring this into the release branch and we're eventually getting rid
of the old dashboard anyway, so I went with a more minimal approach.

Repurposed https://ucsc-cgl.atlassian.net/browse/SEAB-659