Fix Cypress tests
The tooltip was partially over an area being clicked on
and Cypress doesn't allow that by default. I don't know
why this changed with Angular 14, and it's a bit of a hack,
but we click from top to bottom in the search dialog, so
move the tooltip above.

Also noticed an NPE.