Don't requests hits for sidebar ES query
dockstore/dockstore#5257

getSidebarQuery() is an aggregation query (see comment in
search.component.ts two lines up from where it is invoked).

Rename the method to make that clearer.

Set the size of the query to 0. This will still return
aggregations, which is all we use, but it will not return
the actual hits.