Remove ngx-lite/json-ld
dockstore/dockstore#5218

The json-ld code is pretty straightforward, just essentially copied that
code with attribution.

In testing it, I noticed that the json-ld was not showing up on the
home page. I thought I'd broken it, but it turns out that component
is not ever used (json-ld is not showing up in prod either, without my
changes). I deleted the component and moved the json-ld to the
actual home page component.