Don't requests hits for sidebar ES query (#1659)
dockstore/dockstore#5257

Set the size of the query to 0. This will still return
aggregations, which is all we use, but it will not return
the actual hits.

getSidebarQuery() is an aggregation query (see comment in
search.component.ts two lines up from where it is invoked) --
rename the method to make that clearer.