Add ".." at beginning of workerPath (#1712)
Follow up/fix to https://ucsc-cgl.atlassian.net/browse/SEAB-5240

The worker-json.js files were not being downloaded in qa, although they
were in local tests. Because, in a deployment like qa, these files are
served from gui.dockstore.org.

We need the ".." in front of the assets so that build in config.yml
does the string substitution to add gui.dockstore.org.