Download file for app tool fix
dockstore/dockstore#5428

The crux of the issue described in the bug is in
source-file-tabs.service.ts.

In working on it, I noticed compilation errors in the IDE, which led
to a chain of changes.

The starting problem was in workflow.component.html, where
selectedVersion was defined in entry.ts as "WorkflowVersion | Tag", was
being passed to a component that was expecting a WorkflowVersion only.

I moved selectedVersion out of entry.ts and into the workflow and
container subclasses.